Split config.py into more manageable pieces
