+++ frigate/config/__init__.py
@@ -1,0 +1,15 @@
+from frigate.detectors import DetectorConfig, ModelConfig  # noqa: F401
+
+from .auth import *  # noqa: F403
+from .camera import *  # noqa: F403
+from .camera_group import *  # noqa: F403
+from .config import *  # noqa: F403
+from .database import *  # noqa: F403
+from .logger import *  # noqa: F403
+from .mqtt import *  # noqa: F403
+from .notification import *  # noqa: F403
+from .proxy import *  # noqa: F403
+from .semantic_search import *  # noqa: F403
+from .telemetry import *  # noqa: F403
+from .tls import *  # noqa: F403
+from .ui import *  # noqa: F403
