+++ frigate/config/base.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel, ConfigDict
+
+
+class FrigateBaseModel(BaseModel):
+    model_config = ConfigDict(extra="forbid", protected_namespaces=())
