+++ frigate/config/camera/__init__.py
@@ -1,0 +1,17 @@
+from .audio import *  # noqa: F403
+from .birdseye import *  # noqa: F403
+from .camera import *  # noqa: F403
+from .detect import *  # noqa: F403
+from .ffmpeg import *  # noqa: F403
+from .genai import *  # noqa: F403
+from .live import *  # noqa: F403
+from .motion import *  # noqa: F403
+from .mqtt import *  # noqa: F403
+from .objects import *  # noqa: F403
+from .onvif import *  # noqa: F403
+from .record import *  # noqa: F403
+from .review import *  # noqa: F403
+from .snapshots import *  # noqa: F403
+from .timestamp import *  # noqa: F403
+from .ui import *  # noqa: F403
+from .zone import *  # noqa: F403
